대시보드 게시글을 무한히 가져옵니다

+++ app/src/main/java/xyz/parti/catan/models/PartiAccessToken.java
@@ -7,7 +7,7 @@ package xyz.parti.catan.models;
 public class PartiAccessToken {
     public String access_token;
     public String token_type;
-    public Integer expires_in;
+    public Long expires_in;
     public String refresh_token;
 
     public String getValidTokenType() {
